% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{AllObs}
\alias{AllObs}
\title{Display all observations in a table.}
\usage{
AllObs(data = NULL, show = FALSE, label = "Obsn.", within = NULL)
}
\description{
Display all observations in a table.
}
\keyword{internal}
