% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{assert_dependency}
\alias{assert_dependency}
\title{assert if dependency is installed}
\usage{
assert_dependency(library_name, msg = NULL)
}
\description{
assert if dependency is installed
}
\keyword{internal}
