% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_latex.R
\name{clean_latex}
\alias{clean_latex}
\title{Utility function to cleanup LaTeX output from gt and ensures that it
compiles with latex}
\usage{
clean_latex(tab, label = NULL, gof_regex = "^Num Obs.")
}
\arguments{
\item{tab}{table object produced by `modelsummary` or `gt`}

\item{label}{string will be inserted as a `label`}

\item{gof_regex}{regex which identifies the first GOF statistic. Used to figure out
where to insert a midrule to separate coefficients from GOFs.}
}
\value{
a string object with LaTeX code
}
\description{
The `gt::as_latex` function is still in development, rather feature poor,
and prone to breakage. This function is a stopgap measure which adds a
little functionality and "cleans-up" some of the LaTeX output to avoid
common compilation errors. In time, as upstream improves, the goal is to
deprecate this function.
}
