% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{coef_map_felm_iv}
\alias{coef_map_felm_iv}
\title{An unsupported and unexported function to clean the term names from a lfe::felm
instrumental variable model}
\usage{
coef_map_felm_iv(x)
}
\description{
An unsupported and unexported function to clean the term names from a lfe::felm
instrumental variable model
}
\keyword{internal}
