% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{colLabels}
\alias{colLabels}
\title{Retrieve or modify the row or column labels.}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tables}{\code{\link[tables:labels]{colLabels}}}
}}

