% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_modelsummary.R
\name{config_modelsummary}
\alias{config_modelsummary}
\title{Persistent user settings for the \code{modelsummary} package}
\usage{
config_modelsummary(output)
}
\arguments{
\item{output}{Default output format: "kableExtra", "gt", "flextable", "huxtable", "DT", or "markdown"}
}
\description{
Persistent user settings for the \code{modelsummary} package
}
