% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_crosstab.R
\name{datasummary_crosstab}
\alias{datasummary_crosstab}
\title{Cross tabulations for categorical variables}
\usage{
datasummary_crosstab(
  formula,
  statistic = 1 ~ 1 + N + Percent("row"),
  data,
  sparse_header = FALSE,
  ...
)
}
\arguments{
\item{formula}{A two-sided formula to describe the table: rows ~ columns, where
rows and columns are variables in the data. Both rows and column may contain
interactions, e.g., var1 * var2 ~ var3.}

\item{statistic}{A formula of the form 1 ~ 1 + N + Percent("row"). The left-hand
side may only be empty or contain a 1 to include row totals.
The right-hand side may contain: 1 for column totals, N for counts, Percent() for
cell percentages, Percent("row") for row percentages, Percent("col") for column
percentages.}

\item{data}{A data.frame (or tibble)}

\item{sparse_header}{TRUE or FALSE. TRUE eliminates column headers which
have a unique label across all columns, except for the row immediately above
the data. FALSE keeps all headers. The order in which terms are entered in
the formula determines the order in which headers appear.}

\item{...}{passed to datasummary}
}
\value{
a crosstab
}
\description{
Convenience function to tabulate counts, cell percentages, and row/column
percentages for categorical variables.
}
\details{
X
}
\examples{
\dontrun{
}

}
