% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_extract.R
\name{datasummary_extract}
\alias{datasummary_extract}
\title{Extract data from a `tables::tabular` object}
\usage{
datasummary_extract(tab, fmt = NULL, sparse_header = TRUE)
}
\description{
Extract data from a `tables::tabular` object
}
\keyword{internal}
