% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_skim.R
\name{datasummary_skim_numeric}
\alias{datasummary_skim_numeric}
\title{Internal function to skim numeric variables}
\usage{
datasummary_skim_numeric(
  data,
  type = "numeric",
  output = "default",
  fmt = "\%.1f",
  histogram = TRUE,
  title = NULL,
  notes = NULL,
  align = NULL,
  ...
)
}
\description{
Internal function to skim numeric variables
}
\keyword{internal}
