% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.R
\name{escape_html}
\alias{escape_html}
\title{Escape problematic characters to allow display in HTML}
\usage{
escape_html(x)
}
\arguments{
\item{x}{a character string to escape}
}
\description{
Copied from \code{knitr} for internal use because it is unexported and CRAN
rejects :::
}
\keyword{internal}
