% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape problematic characters to allow compilation in LaTeX}
\usage{
escape_latex(x, newlines = FALSE, spaces = FALSE)
}
\arguments{
\item{x}{a character string to escape}

\item{newlines}{boolean}

\item{spaces}{boolean}
}
\description{
Copied from \code{knitr} for internal use because it is unexported and CRAN
rejects :::
}
\keyword{internal}
