% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.R
\name{escape_string}
\alias{escape_string}
\title{Make sure LaTeX and HTML are safe to compile}
\usage{
escape_string(x, output_format = NULL)
}
\description{
Make sure LaTeX and HTML are safe to compile
}
\keyword{internal}
