% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary.R
\name{extract}
\alias{extract}
\title{Extract and combine estimates and goodness-of-fit statistics from several
statistical models.}
\usage{
extract(models, statistic = "std.error", conf_level = 0.95,
  coef_map = NULL, coef_omit = NULL, gof_map = NULL,
  gof_omit = NULL, fmt = "\%.3f", ...)
}
\arguments{
\item{models}{named list of models to summarize}

\item{statistic}{string name of the statistic to include in parentheses
below estimates}

\item{conf_level}{confidence level to use for confidence intervals}

\item{coef_map}{named string vector. Names refer to the original variable
names. Values refer to the variable names that will appear in the table.
Coefficients which are omitted from this vector will be omitted from the
table. The table will be ordered in the same order as this vector.}

\item{coef_omit}{string regular expression. Omits all matching coefficients
from the table (using `stringr::str_detect`).}

\item{gof_map}{data.frame in the same format as `gtsummary::gof_map`}

\item{fmt}{passed to the `sprintf` function to format numeric values into
strings, with rounding, etc. See `?sprintf` for options.}

\item{gof}{string regular expression. Omits all matching gof statistics from
the table (using `stringr::str_detect`).}
}
\description{
Extract and combine estimates and goodness-of-fit statistics from several
statistical models.
}
