% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gof.R
\name{extract_gof}
\alias{extract_gof}
\title{Extract goodness-of-fit statistics from a single model}
\usage{
extract_gof(model, fmt, gof_map, ...)
}
\arguments{
\item{model}{object type with an available `glance` method.}

\item{fmt}{determines how to format numeric values
\itemize{
  \item integer: the number of digits to keep after the period `format(round(x, fmt), nsmall=fmt)`
  \item character: passed to the `sprintf` function (e.g., '\%.3f' keeps 3 digits with trailing zero). See `?sprintf`
  \item function: returns a formatted character string.
}}

\item{gof_map}{\itemize{
  \item NULL (default): the `modelsummary::gof_map` dictionary is used for formatting, and all unknown statistic are included.
  \item data.frame with 3 columns named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples' section below.
  \item list of lists, each of which includes 3 elements named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples section below'.
}}

\item{...}{all other arguments are passed to the `tidy` and `glance` methods
used to extract estimates from the model. For example, this allows users to
set `exponentiate=TRUE` to exponentiate logistic regression coefficients.}
}
\value{
data.frame with goodness-of-fit  statistics
}
\description{
Extract goodness-of-fit statistics from a single model
}
\keyword{internal}
