% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{extract_vcov}
\alias{extract_vcov}
\title{Allow users to override uncertainty estimates}
\usage{
extract_vcov(model, vcov, conf_level = NULL)
}
\arguments{
\item{model}{object type with an available `tidy` method.}

\item{vcov}{robust standard errors and other manual statistics. The `vcov` argument accepts five types of input (see the 'Details' and 'Examples' sections below):
\itemize{
  \item string or list of strings: "robust", "HC", "HC0", "HC1", "HC2", "HC3", "HC4", "HC4m", "HC5", "stata", or "classical" (alias "constant" or "iid").
  \item formula or list of formulas with the cluster variable(s) on the right-hand side (e.g., ~clusterid).
  \item function or list of functions which return variance-covariance matrices with row and column names equal to the names of your coefficient estimates (e.g., `stats::vcov`, `sandwich::vcovHC`).
  \item list of `length(models)` variance-covariance matrices with row and column names equal to the names of your coefficient estimates.
  \item a list of length(models) vectors with names equal to the names of your coefficient estimates. See 'Examples' section below. 
}}

\item{conf_level}{confidence level to use for confidence intervals}
}
\value{
a numeric vector of test statistics
}
\description{
Allow users to override uncertainty estimates
}
\keyword{internal}
