% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{Factor}
\alias{Factor}
\title{Use a variable as a factor to give rows in a table.}
\usage{
Factor(
  x,
  name = deparse(expr),
  levelnames = levels(x),
  texify = getOption("tables.texify", FALSE),
  expr = substitute(x),
  override = TRUE
)
}
\description{
Use a variable as a factor to give rows in a table.
}
\keyword{internal}
