% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{factory}
\alias{factory}
\title{Factory to create tables in different output formats using standardized
inputs.}
\usage{
factory(tab, hrule, notes, output, span, title)
}
\description{
Factory to create tables in different output formats using standardized
inputs.
}
