% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_dataframe.R
\name{factory_dataframe}
\alias{factory_dataframe}
\title{Internal function to build table with `data.frame` with other arguments as
attributes. Useful for testing.}
\usage{
factory_dataframe(
  tab,
  hrule = NULL,
  notes = NULL,
  output_file = NULL,
  output_format = NULL,
  span = NULL,
  title = NULL
)
}
\arguments{
\item{notes}{list or vector of notes to append to the bottom of the table.}

\item{title}{string}
}
\value{
data.frame
}
\description{
Internal function to build table with `data.frame` with other arguments as
attributes. Useful for testing.
}
\keyword{internal}
