% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_flextable.R
\name{factory_flextable}
\alias{factory_flextable}
\title{Internal function to build table with `flextable`}
\usage{
factory_flextable(
  tab,
  align = NULL,
  hrule = NULL,
  notes = NULL,
  output_file = NULL,
  output_format = "flextable",
  span = NULL,
  title = NULL
)
}
\arguments{
\item{tab}{table body (data.frame)}

\item{hrule}{position of horizontal rules (integer vector)}

\item{notes}{list or vector of notes to append to the bottom of the table.}

\item{output_file}{file path (character or NULL)}

\item{output_format}{character}

\item{span}{list of column label spans}

\item{title}{string}
}
\value{
flextable object
}
\description{
Internal function to build table with `flextable`
}
\keyword{internal}
