% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_gt.R
\name{factory_gt}
\alias{factory_gt}
\title{Internal function to build table with `gt`}
\usage{
factory_gt(
  tab,
  hrule = NULL,
  notes = NULL,
  output_file = NULL,
  output_format = "gt",
  span = NULL,
  title = NULL
)
}
\arguments{
\item{tab}{table body (data.frame)}

\item{hrule}{position of horizontal rules (integer vector)}

\item{notes}{list or vector of notes to append to the bottom of the table.}

\item{output_file}{file path (character or NULL)}

\item{output_format}{character}

\item{span}{list of column label spans}

\item{title}{string}
}
\value{
tbl_gt object
}
\description{
Internal function to build table with `gt`
}
\keyword{internal}
