% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_kableExtra.R
\name{factory_kableExtra}
\alias{factory_kableExtra}
\title{Internal function to build table with `gt`}
\usage{
factory_kableExtra(
  tab,
  title = NULL,
  stars = FALSE,
  notes = NULL,
  hrule = NULL,
  span = NULL,
  output_file,
  output_format
)
}
\arguments{
\item{title}{string}

\item{stars}{to indicate statistical significance
\itemize{
  \item FALSE (default): no significance stars. 
  \item TRUE: *=.1, **=.05, ***=.01
  \item Named numeric vector for custom stars such as `c('*' = .1, '+' = .05)`
}}

\item{notes}{list or vector of notes to append to the bottom of the table.}
}
\value{
tbl_gt object
}
\description{
Internal function to build table with `gt`
}
\keyword{internal}
