% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_factory.R
\name{fmt_decimal}
\alias{fmt_decimal}
\title{Rounding with decimal digits in the \code{fmt} argument}
\usage{
fmt_decimal(digits = 3, pdigits = NULL, ...)
}
\arguments{
\item{digits}{Number of decimal digits to keep, including trailing zeros.}

\item{pdigits}{Number of decimal digits to keep for p values. If \code{NULL}, the value of \code{digits} is used.}

\item{...}{Additional arguments are passed to the \code{format()} function (e.g., \code{big.marks}, \code{scientific}). See \code{?format}}
}
\description{
Rounding with decimal digits in the \code{fmt} argument
}
