% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_factory.R
\name{fmt_function}
\alias{fmt_function}
\title{Rounding with a user-supplied function in the \code{fmt} argument}
\usage{
fmt_function(fun)
}
\arguments{
\item{fun}{A function which accepts a numeric vector and returns a numeric vector of the same length.}
}
\description{
Rounding with a user-supplied function in the \code{fmt} argument
}
