% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{fmt_labels_md}
\alias{fmt_labels_md}
\title{A convenience function to render markdown to html in row and column labels}
\usage{
fmt_labels_md(tab, position = c("both", "row", "column"))
}
\arguments{
\item{tab}{a `gt` table object}

\item{position}{character string determines wither row, column or both
labels should be rendered.}
}
\description{
A convenience function to render markdown to html in row and column labels
}
\note{
This function only works for HTML output, since the `gt` render tools
are less developed for LaTeX and RTF output.
}
\keyword{internal}
