% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_factory.R
\name{fmt_scientific}
\alias{fmt_scientific}
\title{Rounding using scientific notation}
\usage{
fmt_scientific(digits = 3, ...)
}
\arguments{
\item{digits}{a positive integer indicating how many significant digits are to be used for numeric and complex \code{x}.}

\item{...}{additional arguments passed to \code{format()}.}
}
\description{
Rounding using scientific notation
}
