% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_factory.R
\name{fmt_significant}
\alias{fmt_significant}
\title{Rounding with significant digits in the \code{fmt} argument}
\usage{
fmt_significant(digits = 3, ...)
}
\arguments{
\item{digits}{Number of significant digits to keep.}

\item{...}{Additional arguments are passed to the \code{format()} function (e.g., \code{big.marks}, \code{scientific}). See \code{?format}}
}
\description{
The number of decimal digits to keep after the decimal is assessed
}
