% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_factory.R
\name{fmt_sprintf}
\alias{fmt_sprintf}
\title{Rounding with the \code{sprintf()} function in the \code{fmt} argument}
\usage{
fmt_sprintf(fmt)
}
\arguments{
\item{fmt}{A string to control \code{sprintf()}, such as \code{"\%.3f"} to keep 3 decimal digits. See \code{?sprintf}}
}
\description{
Rounding with the \code{sprintf()} function in the \code{fmt} argument
}
