% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_factory.R
\name{fmt_term}
\alias{fmt_term}
\title{Rounding with decimal digits on a per-term basis in the \code{fmt} argument for \code{modelsummary()}}
\usage{
fmt_term(..., default = 3)
}
\arguments{
\item{...}{Term names and \code{fmt} value}

\item{default}{Number of decimal digits to keep for unspecified terms}
}
\description{
Rounding with decimal digits on a per-term basis in the \code{fmt} argument for \code{modelsummary()}
}
