% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gof.R
\name{get_gof_parameters}
\alias{get_gof_parameters}
\title{Extract goodness-of-fit statistics from a single model using
the \code{performance} package}
\usage{
get_gof_parameters(model, ...)
}
\description{
Extract goodness-of-fit statistics from a single model using
the \code{performance} package
}
\keyword{internal}
