% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gof.R
\name{get_gof_performance}
\alias{get_gof_performance}
\title{Extract goodness-of-fit statistics from a single model using
the `performance` package}
\usage{
get_gof_performance(model, ...)
}
\description{
Extract goodness-of-fit statistics from a single model using
the `performance` package
}
\keyword{internal}
