% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{get_variable_labels_data}
\alias{get_variable_labels_data}
\title{Return labels from data frame}
\usage{
get_variable_labels_data(data)
}
\description{
Return labels from data frame
}
\examples{
\dontrun{
}
}
