% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_labels.R
\name{get_variable_labels_models}
\alias{get_variable_labels_models}
\title{Create a named vector of variable labels from a list of models}
\usage{
get_variable_labels_models(models)
}
\arguments{
\item{models}{A list of models}
}
\value{
A named vector of variable labels
}
\description{
Create a named vector of variable labels from a list of models
}
