% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcov.R
\name{get_vcov_type}
\alias{get_vcov_type}
\title{internal function}
\usage{
get_vcov_type(vcov)
}
\arguments{
\item{v}{a string or formula describing the standard error type}
}
\description{
internal function
}
\keyword{internal}
