% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{glance.fixest}
\alias{glance.fixest}
\title{Glance a `fixest` model}
\usage{
\method{glance}{fixest}(x, ...)
}
\arguments{
\item{x}{A model object produced by the `fixest::feols` or `fixest::feglm` functions}

\item{...}{extra arguments (not used)}
}
\value{
a dataframe with one row
}
\description{
Glance a `fixest` model
}
\seealso{
Other tidiers: 
\code{\link{glance.MIresult}()},
\code{\link{glance.mira}()},
\code{\link{tidy.fixest}()}
}
\concept{tidiers}
\keyword{internal}
