% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{glance.MIresult}
\alias{glance.MIresult}
\title{Glance a multiple imputation `mitools::MIcombine` pooled object}
\usage{
\method{glance}{MIresult}(x, ...)
}
\arguments{
\item{x}{An object with multiply-imputed models from `mice` (class: `mira`)}

\item{...}{extra arguments (not used)}
}
\value{
a dataframe with one row
}
\description{
Glance a multiple imputation `mitools::MIcombine` pooled object
}
\seealso{
Other tidiers: 
\code{\link{glance.mira}()}
}
\concept{tidiers}
\keyword{internal}
