% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsummary_wide.R
\name{glance.modelsummary_internal_coef_groups}
\alias{glance.modelsummary_internal_coef_groups}
\title{internal glance method}
\usage{
\method{glance}{modelsummary_internal_coef_groups}(x)
}
\description{
internal glance method
}
\keyword{internal}
