% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_custom.R
\name{glance_custom}
\alias{glance_custom}
\title{Extract custom information from a model object and turn it into a tidy
data.frame or tibble with a single row.}
\usage{
glance_custom(x)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}
}
\description{
glance_custom methods always return either a one-row data frame (except on
 `NULL`, which returns an empty data frame). This
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("glance")}
}

