% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_custom.R
\name{glance_custom_internal.fixest}
\alias{glance_custom_internal.fixest}
\title{Avoid namespace conflict when we want to customize glance internally and
still allow users to do the same with their own functions}
\usage{
\method{glance_custom_internal}{fixest}(x, vcov_type = NULL, ...)
}
\description{
Avoid namespace conflict when we want to customize glance internally and
still allow users to do the same with their own functions
}
\keyword{internal}
