% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_stats.R
\name{glance_custom_internal.lm}
\alias{glance_custom_internal.lm}
\title{Avoid namespace conflict when we want to customize glance internally and
still allow users to do the same with their own functions}
\usage{
\method{glance_custom_internal}{lm}(x, vcov_type = NULL, gof = NULL, ...)
}
\description{
Avoid namespace conflict when we want to customize glance internally and
still allow users to do the same with their own functions
}
\keyword{internal}
