% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats.R
\name{glance_easystats}
\alias{glance_easystats}
\title{extract gof using the easystats `performance` package}
\usage{
glance_easystats(model, ...)
}
\description{
extract gof using the easystats `performance` package
}
\keyword{internal}
