% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_extra.R
\name{glance_more}
\alias{glance_more}
\title{Glance *more* at an object}
\usage{
glance_more(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\description{
Construct a single row summary "glance" of a model, fit, or other object.
This method extracts more information than the usual `glance` method.
Defining a `glance_more` allows users to customize the features of a model
to display in a `modelsummary` table.
}
\details{
glance_more methods must always return a one-row data frame
}
