% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary.R
\docType{data}
\name{gof_map}
\alias{gof_map}
\alias{sanity_check}
\title{Data.frame used to clean up and format goodness-of-fit statistics}
\format{data.frame with 3 columns of character data: raw, clean, fmt
Internal function to check the sanity of various user inputs}
\usage{
sanity_check(gof_map = NULL)
}
\description{
Data.frame used to clean up and format goodness-of-fit statistics
}
\keyword{datasets}
