% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary.R
\name{gtsummary}
\alias{gtsummary}
\title{Beautiful, customizable summaries of statistical models}
\usage{
gtsummary(models, statistic = "std.error", conf_level = 0.95,
  coef_map = NULL, coef_omit = NULL, gof_map = NULL,
  gof_omit = NULL, fmt = "\%.3f", title = NULL, subtitle = NULL,
  filename = NULL, ...)
}
\description{
Beautiful, customizable summaries of statistical models
}
