% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{Heading}
\alias{Heading}
\title{'Heading' pseudo-function}
\usage{
Heading(name = NULL, override = TRUE, character.only = FALSE, nearData = TRUE)
}
\description{
'Heading' pseudo-function
}
\keyword{internal}
