% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_functions.R
\name{Histogram}
\alias{Histogram}
\title{datasummary statistic shortcut}
\usage{
Histogram(x, bins = 10)
}
\arguments{
\item{x}{varible to summarize}

\item{bins}{number of histogram bars}
}
\description{
This function uses Unicode characters to create a histogram. This can
sometimes be useful, but is generally discouraged. Unicode characters can
only display a limited number of heights for bars, and the accuracy of
output is highly dependent on the platform (typeface, output type, windows
vs. mac, etc.). We recommend you use the `kableExtra::spec_hist` function
instead.
}
