% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_estimates.R
\name{map_estimates}
\alias{map_estimates}
\title{rename and reorder estimates from a \emph{single} model
(before merging to collapse)}
\usage{
map_estimates(
  estimates,
  coef_rename,
  coef_map,
  coef_omit,
  group_map,
  shape = NULL
)
}
\description{
rename and reorder estimates from a \emph{single} model
(before merging to collapse)
}
\keyword{internal}
