% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsummary.R
\name{map_omit_gof}
\alias{map_omit_gof}
\title{Internal function to subset, rename and re-order gof statistics}
\usage{
map_omit_gof(gof, gof_omit, gof_map)
}
\description{
Internal function to subset, rename and re-order gof statistics
}
\keyword{internal}
