% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsummary.R
\name{map_omit_rename_estimates}
\alias{map_omit_rename_estimates}
\title{rename and reorder estimates from a \emph{single} model
(before merging to collapse)}
\usage{
map_omit_rename_estimates(
  estimates,
  coef_rename,
  coef_map,
  coef_omit,
  group_map
)
}
\description{
rename and reorder estimates from a \emph{single} model
(before merging to collapse)
}
\keyword{internal}
