% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{Multicolumn}
\alias{Multicolumn}
\title{Use a variable as a factor to give rows in a table.}
\usage{
Multicolumn(
  x,
  name = deparse(expr),
  levelnames = levels(x),
  width = 2,
  first = 1,
  justify = "l",
  texify = getOption("tables.texify", FALSE),
  expr = substitute(x),
  override = TRUE
)
}
\description{
Use a variable as a factor to give rows in a table.
}
\keyword{internal}
