% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_functions.R
\name{N}
\alias{N}
\title{datasummary statistic shortcut}
\usage{
N(x)
}
\arguments{
\item{x}{variable to summarize}
}
\description{
datasummary statistic shortcut
}
\examples{
\dontrun{
datasummary(Factor(cyl) ~ N, data = mtcars)
}
}
