% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_functions.R
\name{Ncol}
\alias{Ncol}
\title{datasummary statistic shortcut}
\usage{
Ncol(x, ...)
}
\arguments{
\item{x}{variable to summarize}

\item{...}{unused}
}
\description{
datasummary statistic shortcut
}
