% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_functions.R
\name{NPercent}
\alias{NPercent}
\title{datasummary statistic shortcut}
\usage{
NPercent(x, y)
}
\arguments{
\item{x}{variable to summarize}

\item{y}{denominator variable}
}
\description{
datasummary statistic shortcut
}
\keyword{internal}
