% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_functions.R
\name{NUnique}
\alias{NUnique}
\title{datasummary statistic shortcut}
\usage{
NUnique(x, ...)
}
\arguments{
\item{x}{variable to summarize}

\item{...}{unused}
}
\description{
datasummary statistic shortcut
}
\examples{
\dontrun{
datasummary(cyl + hp ~ NUnique, data = mtcars)
}
}
\keyword{internal}
