% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{override_statistic.character}
\alias{override_statistic.character}
\title{Use the statistic_override vector to extract std.error/p.value/statistic}
\usage{
\method{override_statistic}{character}(x, ...)
}
\arguments{
\item{...}{all other arguments are passed to the `tidy` method used to
extract estimates from the model. For example, this allows users to set
`exponentiate=TRUE` to exponentiate logistic regression coefficients.}

\item{model}{object type with an available `tidy` method.}
}
\value{
tibbl
}
\description{
Use the statistic_override vector to extract std.error/p.value/statistic
}
\keyword{internal}
