% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{override_statistic.function}
\alias{override_statistic.function}
\title{Use the statistic_override function to extract std.error}
\usage{
\method{override_statistic}{`function`}(x, model, statistic, ...)
}
\arguments{
\item{model}{object type with an available `tidy` method.}

\item{statistic}{string name of the statistic to include in parentheses
\itemize{
  \item Typical values: "conf.int", "std.error", "statistic", "p.value"
  \item Alternative values: any column name produced by `broom::tidy(model)`
}}

\item{...}{all other arguments are passed to the `tidy` method used to
extract estimates from the model. For example, this allows users to set
`exponentiate=TRUE` to exponentiate logistic regression coefficients.}
}
\value{
tibble
}
\description{
Use the statistic_override function to extract std.error
}
\keyword{internal}
