% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{override_statistic.matrix}
\alias{override_statistic.matrix}
\title{Use the statistic_override vector to extract std.error/p.value/statistic}
\usage{
\method{override_statistic}{matrix}(x, statistic, ...)
}
\arguments{
\item{statistic}{string name of the statistic to include in parentheses
\itemize{
  \item Typical values: "conf.int", "std.error", "statistic", "p.value"
  \item Alternative values: any column name produced by `broom::tidy(model)`
}}

\item{...}{all other arguments are passed to the `tidy` method used to
extract estimates from the model. For example, this allows users to set
`exponentiate=TRUE` to exponentiate logistic regression coefficients.}

\item{model}{object type with an available `tidy` method.}
}
\value{
tibbl
}
\description{
Use the statistic_override vector to extract std.error/p.value/statistic
}
\keyword{internal}
