% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{override_statistic}
\alias{override_statistic}
\title{Internal method generic to extract statistic override}
\usage{
override_statistic(x, ...)
}
\description{
Internal method generic to extract statistic override
}
\keyword{internal}
