% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pad.R
\name{pad}
\alias{pad}
\title{pad a character vector with spaces}
\usage{
pad(x)
}
\description{
if there are duplicate or empty entries in a vector, this function adds
an arbitrary number of empty spaces to each entry to de-duplicate. this
is useful for table-making packages which require unique colnames. they
will usually strip white space anyway. This is suboptimal because it
makes it harder to select columns, so we only apply pad duplicate
entries.
}
\keyword{internal}
