% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output_arg.R
\name{parse_output_arg}
\alias{parse_output_arg}
\title{parse an `output` argument to determine which table factory and which output
format to use.}
\usage{
parse_output_arg(output)
}
\description{
parse an `output` argument to determine which table factory and which output
format to use.
}
\keyword{internal}
