% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{Paste}
\alias{Paste}
\title{Generate terms to paste values together in table.}
\usage{
Paste(
  ...,
  head,
  digits = 2,
  justify = "c",
  prefix = "",
  sep = "",
  postfix = "",
  character.only = FALSE
)
}
\description{
Generate terms to paste values together in table.
}
\keyword{internal}
