% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{Percent}
\alias{Percent}
\title{Pseudo-function to compute a statistic relative to a reference set.}
\usage{
Percent(denom = "all", fn = percent)
}
\description{
Pseudo-function to compute a statistic relative to a reference set.
}
\keyword{internal}
