% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_functions.R
\name{PercentMissing}
\alias{PercentMissing}
\title{datasummary statistic shortcut}
\usage{
PercentMissing(x)
}
\arguments{
\item{x}{variable to summarize}
}
\description{
datasummary statistic shortcut
}
\keyword{internal}
