% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{PlusMinus}
\alias{PlusMinus}
\title{Generate 'x +/- y' terms in table.}
\usage{
PlusMinus(x, y, head, xhead, yhead, digits = 2, character.only = FALSE, ...)
}
\description{
Generate 'x +/- y' terms in table.
}
\keyword{internal}
