% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R, R/utils_pipe.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

