% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_replace.R
\name{replace_dict}
\alias{replace_dict}
\title{replace character vector entries using a named vector dictionary}
\usage{
replace_dict(x, dict)
}
\description{
replace character vector entries using a named vector dictionary
}
\keyword{internal}
