% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rounding.R
\name{rounding}
\alias{rounding}
\title{Convert numeric values to strings using the `sprintf` function. NA, NaN,
-Inf, and Inf are replaced by an empty string.}
\usage{
rounding(x, fmt = "\%.3f")
}
\arguments{
\item{x}{a numeric vector to be converted to string}

\item{fmt}{a character vector of format strings which will be fed to the
`sprintf` function. See ?sprintf for details.}
}
\value{
a rounded number as character
}
\description{
Convert numeric values to strings using the `sprintf` function. NA, NaN,
-Inf, and Inf are replaced by an empty string.
}
\keyword{internal}
