% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{RowFactor}
\alias{RowFactor}
\title{Use a variable as a factor to give rows in a table.}
\usage{
RowFactor(
  x,
  name = deparse(expr),
  levelnames = levels(x),
  spacing = 3,
  space = 1,
  suppressfirst = TRUE,
  nopagebreak = "\\\\nopagebreak ",
  texify = getOption("tables.texify", FALSE),
  expr = substitute(x),
  override = TRUE
)
}
\description{
Use a variable as a factor to give rows in a table.
}
\keyword{internal}
