% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{RowNum}
\alias{RowNum}
\title{Display all observations in a table.}
\usage{
RowNum(within = NULL, perrow = 5, show = FALSE, label = "Row", data = NULL)
}
\description{
Display all observations in a table.
}
\keyword{internal}
