% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{sanity_checks}
\alias{sanity_checks}
\title{internal function to check the sanity of user input}
\usage{
sanity_checks(models, statistic = "std.error",
  statistic_override = NULL, conf_level = 0.95, coef_map = NULL,
  coef_omit = NULL, gof_map = NULL, gof_omit = NULL, fmt = "\%.3f",
  stars = NULL, stars_note = TRUE, title = NULL, subtitle = NULL,
  notes = NULL, add_rows = NULL, filename = NULL)
}
\description{
internal function to check the sanity of user input
}
