% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{sanity_ds_nesting_factor}
\alias{sanity_ds_nesting_factor}
\title{sanity check: datasummary}
\usage{
sanity_ds_nesting_factor(formula, data)
}
\description{
sanity check: datasummary
}
\keyword{internal}
