% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{sanity_ds_right_handed_formula}
\alias{sanity_ds_right_handed_formula}
\title{sanity check: datasummary_table1}
\usage{
sanity_ds_right_handed_formula(formula)
}
\description{
right-handed formulae only
}
