% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{statistic_override_function}
\alias{statistic_override_function}
\title{Use the statistic_override function to extract std.error}
\usage{
statistic_override_function(model, statistic_override)
}
\description{
Use the statistic_override function to extract std.error
}
