% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{statistic_override_lmtest}
\alias{statistic_override_lmtest}
\title{Use the lmtest::coeftest function to extract uncertainty estimates}
\usage{
statistic_override_lmtest(model, statistic_override)
}
\description{
Use the lmtest::coeftest function to extract uncertainty estimates
}
