% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{statistic_override_matrix}
\alias{statistic_override_matrix}
\title{Use the statistic_override matrix to extract std.error}
\usage{
statistic_override_matrix(model, statistic_override)
}
\description{
Use the statistic_override matrix to extract std.error
}
