% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{statistic_override_vector}
\alias{statistic_override_vector}
\title{Use the statistic_override vector to extract std.error/p.value/statistic}
\usage{
statistic_override_vector(model, statistic_override, statistic)
}
\description{
Use the statistic_override vector to extract std.error/p.value/statistic
}
