% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_models.R
\name{supported_models}
\alias{supported_models}
\title{List of model objects from which \code{modelsummary} can extract estimates and
statistics}
\usage{
supported_models()
}
\description{
List of model objects from which \code{modelsummary} can extract estimates and
statistics
}
\keyword{internal}
