% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.fixest}
\alias{tidy.fixest}
\title{Tidy a `fixest` model}
\usage{
\method{tidy}{fixest}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A model object produced by the `fixest::feols` or `fixest::feglm` functions}

\item{...}{extra arguments (not used)}
}
\value{
a dataframe with one row per term
}
\description{
Tidy a `fixest` model
}
\seealso{
Other tidiers: 
\code{\link{glance.MIresult}()},
\code{\link{glance.fixest}()},
\code{\link{glance.mira}()}
}
\concept{tidiers}
\keyword{internal}
