% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.MIresult}
\alias{tidy.MIresult}
\title{Tidy multiple imputation models created with `mitools::MIcombine`}
\usage{
\method{tidy}{MIresult}(x, ...)
}
\arguments{
\item{x}{A `mira` object containing multiple models based on `mice` imputations.}

\item{...}{extra arguments (not used)}
}
\value{
a dataframe with one row per term
}
\description{
Tidy multiple imputation models created with `mitools::MIcombine`
}
\keyword{internal}
