% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsummary_wide.R
\name{tidy.modelsummary_internal_coef_groups}
\alias{tidy.modelsummary_internal_coef_groups}
\title{internal tidy method}
\usage{
\method{tidy}{modelsummary_internal_coef_groups}(x)
}
\description{
internal tidy method
}
\keyword{internal}
