% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom.R
\name{tidy_custom}
\alias{tidy_custom}
\title{Extract custom information from a model object and turn it into a tidy
data.frame or tibble}
\usage{
tidy_custom(x)
}
\arguments{
\item{x}{An object to be converted into a tidy data.frame or tibble.}
}
\value{
A data.frame or tibble with information about model components.
}
\description{
Extract custom information from a model object and turn it into a tidy
data.frame or tibble
}
\keyword{internal}
