% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom.R
\name{tidy_custom_internal}
\alias{tidy_custom_internal}
\title{Avoid namespace conflict when we want to customize glance internally and
still allow users to do the same with their own functions}
\usage{
tidy_custom_internal(x, ...)
}
\description{
Avoid namespace conflict when we want to customize glance internally and
still allow users to do the same with their own functions
}
\keyword{internal}
