% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats.R
\name{tidy_easystats}
\alias{tidy_easystats}
\title{extract estimates using the easystats `parameters` package}
\usage{
tidy_easystats(model, ...)
}
\description{
extract estimates using the easystats `parameters` package
}
\keyword{internal}
