% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{tidy_felm_iv}
\alias{tidy_felm_iv}
\title{An unsupported and unexported function to tidy a lfe::felm instrumental
variable model}
\usage{
tidy_felm_iv(x, ...)
}
\description{
An unsupported and unexported function to tidy a lfe::felm instrumental
variable model
}
\keyword{internal}
