% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_modelsummary.R
\name{update_modelsummary}
\alias{update_modelsummary}
\title{Update \code{modelsummary} and its dependencies}
\usage{
update_modelsummary(source = "development")
}
\arguments{
\item{source}{one of two strings: "development" or "cran"}
}
\description{
Update \code{modelsummary} and its dependencies to the latest R-Universe or CRAN versions. The R session needs to be restarted after install.
}
