% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{AllObs}
\alias{AllObs}
\title{Display all observations in a table.}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tables}{\code{\link[tables]{AllObs}}}
}}

