% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_latex.R
\name{clean_latex}
\alias{clean_latex}
\title{Deprecated function}
\usage{
clean_latex(...)
}
\arguments{
\item{...}{catch everything}
}
\description{
The `gt::as_latex` function does not (yet) produce compilable LaTeX code.
This function used to clean up LaTeX output to allow compilation to PDF.
This function is now deprecated since `modelsummary` currently supports
`kableExtra`, which has a mature LaTeX rendering engine.
}
