% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_rename.R
\name{coef_rename}
\alias{coef_rename}
\title{Rename model terms}
\usage{
coef_rename(
  x,
  factor = TRUE,
  factor_name = TRUE,
  backticks = TRUE,
  titlecase = TRUE,
  underscore = TRUE,
  asis = TRUE
)
}
\arguments{
\item{x}{character vector of term names to transform}

\item{factor}{boolean remove the "factor()" label}

\item{factor_name}{boolean remove the "factor()" label and the name of the
variable}

\item{backticks}{boolean remove backticks}

\item{titlecase}{boolean convert to title case}

\item{underscore}{boolean replace underscores by spaces}

\item{asis}{boolean remove the \code{I} from as-is formula calls}
}
\description{
A convenience function which can be passed to the \code{coef_rename} argument of
the \code{modelsummary} function.
}
\examples{
\dontrun{
library(modelsummary)
dat <- mtcars
dat$horse_power <- dat$hp
mod <- lm(mpg ~ horse_power + factor(cyl), dat)
modelsummary(mod, coef_rename = coef_rename)
}
}
