% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary.R
\name{datasummary}
\alias{datasummary}
\title{A thin wrapper around `tables::tabular`}
\usage{
datasummary(
  formula,
  data,
  output = "default",
  title = NULL,
  notes = NULL,
  sparse_header = TRUE,
  ...
)
}
\arguments{
\item{formula}{A two-sided formula which describes a 'tabular' object.
Details can be found in the 'tables' package documentation:
?tables::tabular}

\item{data}{A data.frame (or tibble)}

\item{output}{filename or object type (string)
\itemize{
  \item Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg. 
  \item Supported object types: "default", "html", "markdown", "latex", "gt", "kableExtra", "huxtable", "flextable".
  \item When a file name is supplied to the `output` argument, the table is written immediately to file. If you want to customize your table by post-processing it with functions provided by the `gt` or `kableExtra` packages, you need to choose a different output format (e.g., "gt", "latex", "html", "markdown"), and you need to save the table after post-processing using the `gt::gtsave`, `kable::save_kable`, or `cat` functions.
}}

\item{title}{string}

\item{notes}{list or vector of notes to append to the bottom of the table.}

\item{sparse_header}{TRUE or FALSE
\itemize{
\item TRUE eliminates column headers which have a unique label across all
columns, except for the row immediately above the data.
\item FALSE keeps all headers.
\item Note that the order in which terms are entered in the formula
determines the order in which headers appear. For example, `x~mean*z` will
print the `mean`-related header above the `z`-related header.`
}}

\item{...}{all other arguments are passed to the `tidy` method used to
extract estimates from the model. For example, this allows users to set
`exponentiate=TRUE` to exponentiate logistic regression coefficients.}
}
\description{
A thin wrapper around `tables::tabular`
}
