% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_correlation.R
\name{datasummary_correlation}
\alias{datasummary_correlation}
\title{datasummary template to create a correlation table}
\usage{
datasummary_correlation(
  data,
  output = "default",
  title = NULL,
  notes = NULL,
  ...
)
}
\description{
datasummary template to create a correlation table
}
