% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_skim.R
\name{datasummary_skim}
\alias{datasummary_skim}
\title{datasummary template}
\usage{
datasummary_skim(
  data,
  type = "numeric",
  output = "default",
  fmt = "\%.1f",
  title = NULL,
  notes = NULL,
  histogram = FALSE,
  ...
)
}
\description{
datasummary template
}
