% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary_skim.R
\name{datasummary_skim_categorical}
\alias{datasummary_skim_categorical}
\title{Internal function to skim categorical variables}
\usage{
datasummary_skim_categorical(data, output, fmt, title, notes, align, ...)
}
\description{
Internal function to skim categorical variables
}
\keyword{internal}
