% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary.R
\name{extract_gof}
\alias{extract_gof}
\title{Extract goodness-of-fit statistics from a single model}
\usage{
extract_gof(model, fmt = "\%.3f", gof_map = NULL, ...)
}
\description{
Extract goodness-of-fit statistics from a single model
}
