% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{extract_statistic_override}
\alias{extract_statistic_override}
\title{Allow users to override uncertainty estimates
importFrom broom tidy}
\usage{
extract_statistic_override(model, statistic_override,
  statistic = "std.error")
}
\description{
Allow users to override uncertainty estimates
importFrom broom tidy
}
