% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{extract_statistic_override}
\alias{extract_statistic_override}
\title{Allow users to override uncertainty estimates}
\usage{
extract_statistic_override(model, statistic_override, conf_level = NULL)
}
\arguments{
\item{model}{object type with an available `tidy` method.}

\item{statistic_override}{deprecated argument. Use the `estimate_override` argument instead.}

\item{conf_level}{confidence level to use for confidence intervals}
}
\value{
a numeric vector of test statistics
}
\description{
Allow users to override uncertainty estimates
}
\keyword{internal}
