% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_flextable.R
\name{factory_flextable}
\alias{factory_flextable}
\title{Internal function to build table with `flextable`}
\usage{
factory_flextable(
  tab,
  hrule = NULL,
  notes = NULL,
  output_file = NULL,
  output_format = "flextable",
  span = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{notes}{list or vector of notes to append to the bottom of the table.}

\item{title}{string}

\item{...}{all other arguments are passed to the `tidy` method used to
extract estimates from the model. For example, this allows users to set
`exponentiate=TRUE` to exponentiate logistic regression coefficients.}
}
\value{
flextable object
}
\description{
Internal function to build table with `flextable`
}
\keyword{internal}
