% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_gt.R
\name{factory_gt}
\alias{factory_gt}
\title{Internal function to build table with `gt`}
\usage{
factory_gt(
  tab,
  title = NULL,
  stars = FALSE,
  notes = NULL,
  hrule = NULL,
  span = NULL,
  output_file,
  output_format,
  ...
)
}
\arguments{
\item{title}{string}

\item{stars}{to indicate statistical significance
\itemize{
  \item FALSE (default): no significance stars. 
  \item TRUE: *=.1, **=.05, ***=.01
  \item Named numeric vector for custom stars such as `c('*' = .1, '+' = .05)`
}}

\item{notes}{list or vector of notes to append to the bottom of the table.}

\item{...}{all other arguments are passed to the `tidy` method used to
extract estimates from the model. For example, this allows users to set
`exponentiate=TRUE` to exponentiate logistic regression coefficients.}

\item{stars_note}{argument passed by `modelsummary()`}
}
\value{
tbl_gt object
}
\description{
Internal function to build table with `gt`
}
\keyword{internal}
