% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_gt.R
\name{factory_gt}
\alias{factory_gt}
\title{Internal function to build table with `gt`}
\usage{
factory_gt(
  tab,
  hrule = NULL,
  notes = NULL,
  output_file = NULL,
  output_format = "gt",
  span = NULL,
  title = NULL,
  ...
)
}
\value{
tbl_gt object
}
\description{
Internal function to build table with `gt`
}
