% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_huxtable.R
\name{factory_huxtable}
\alias{factory_huxtable}
\title{Internal function to build table with `huxtable`}
\usage{
factory_huxtable(
  tab,
  title,
  stars,
  notes,
  hrule,
  output_file,
  output_format,
  ...
)
}
\arguments{
\item{title}{string}

\item{stars}{to indicate statistical significance
\itemize{
  \item FALSE (default): no significance stars. 
  \item TRUE: *=.1, **=.05, ***=.01
  \item Named numeric vector for custom stars such as `c('*' = .1, '+' = .05)`
}}

\item{notes}{list or vector of notes to append to the bottom of the table.}

\item{...}{all other arguments are passed to the `tidy` method used to
extract estimates from the model. For example, this allows users to set
`exponentiate=TRUE` to exponentiate logistic regression coefficients.}

\item{stars_note}{passed by `modelsummary()`}
}
\value{
huxtable object
}
\description{
Internal function to build table with `huxtable`
}
\keyword{internal}
