% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_kableExtra.R
\name{factory_kableExtra}
\alias{factory_kableExtra}
\title{Internal function to build table with `gt`}
\usage{
factory_kableExtra(
  tab,
  hrule = NULL,
  notes = NULL,
  output_file = NULL,
  output_format = "kableExtra",
  span = span,
  title = NULL,
  ...
)
}
\arguments{
\item{notes}{list or vector of notes to append to the bottom of the table.}

\item{title}{string}

\item{...}{all other arguments are passed to the `tidy` method used to
extract estimates from the model. For example, this allows users to set
`exponentiate=TRUE` to exponentiate logistic regression coefficients.}
}
\value{
tbl_gt object
}
\description{
Internal function to build table with `gt`
}
\keyword{internal}
