% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Extract model estimates. A mostly internal function with some potential uses
outside.}
\usage{
get_estimates(model, conf_level = 0.95, ...)
}
\arguments{
\item{model}{a single model object}

\item{conf_level}{confidence level to use for confidence intervals}

\item{...}{all other arguments are passed through to the extractor and
table-making functions. This allows users to pass arguments directly to
\code{modelsummary} in order to affect the behavior of other functions behind
the scenes. Examples include:
\itemize{
\item \code{broom::tidy(exponentiate=TRUE)} to exponentiate logistic regression
\item \code{kableExtra::kbl(escape=FALSE)} to avoid escaping math characters in #'   \code{kableExtra} tables.
\item \code{performance::model_performance(metrics="RMSE")} to select goodness-of-fit #'   statistics to extract using the \code{performance} package (must have set #'   \code{options(modelsummary_get="easystats")} first).
}}
}
\description{
Extract model estimates. A mostly internal function with some potential uses
outside.
}
