% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Extract model estimates. A mostly internal function with some potential uses
outside.}
\usage{
get_estimates(model, conf_level = 0.95, ...)
}
\arguments{
\item{model}{a single model object}

\item{conf_level}{confidence level to use for confidence intervals}

\item{...}{all other arguments are passed through to the extractor and
table-making functions. This allows users to specify additional options such
as set `broom::tidy(exponentiate=TRUE)` to exponentiate logistic regression
coefficients or `kableExtra::kbl(escape=FALSE)` to avoid escaping math
characters in `kableExtra` tables.}
}
\description{
Extract model estimates. A mostly internal function with some potential uses
outside.
}
