% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Extract model estimates. A mostly internal function with some potential uses
outside.}
\usage{
get_estimates(model, conf_level = 0.95, ...)
}
\arguments{
\item{model}{a single model object}

\item{conf_level}{confidence level to use for confidence intervals}

\item{...}{all other arguments are passed to the `tidy` and `glance` methods
used to extract estimates from the model. For example, this allows users to
set `exponentiate=TRUE` to exponentiate logistic regression coefficients.}
}
\description{
Extract model estimates. A mostly internal function with some potential uses
outside.
}
