% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gof.R
\name{get_gof}
\alias{get_gof}
\title{Extract model gof A mostly internal function with some potential uses
outside.}
\usage{
get_gof(model, vcov_type = NULL, ...)
}
\arguments{
\item{model}{a single model object}

\item{vcov_type}{string vcov type to add at the bottom of the table}

\item{...}{all other arguments are passed through to three functions. See the documentation of these functions for lists of available arguments.
\itemize{
\item \link[parameters:model_parameters]{parameters::model_parameters} extracts parameter estimates. Available arguments depend on model type, but include:
\itemize{
\item \code{standardize}, \code{centrality}, \code{dispersion}, \code{test}, \code{ci_method}, \code{prior}, \code{diagnostic}, \code{rope_range}, \code{power}, \code{cluster}, etc.
}
\item \link[performance:model_performance]{performance::model_performance} extracts goodness-of-fit statistics. Available arguments depend on model type, but include:
\itemize{
\item \code{metrics}, \code{estimator}, etc.
}
\item \link[kableExtra:kbl]{kableExtra::kbl} or \link[gt:gt]{gt::gt} draw tables, depending on the value of the \code{output} argument.
}}
}
\description{
Extract model gof A mostly internal function with some potential uses
outside.
}
