% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gof.R
\name{get_gof}
\alias{get_gof}
\title{Extract model gof A mostly internal function with some potential uses
outside.}
\usage{
get_gof(model, ...)
}
\arguments{
\item{model}{a single model object}

\item{...}{all other arguments are passed to the `tidy` and `glance` methods
used to extract estimates from the model. For example, this allows users to
set `exponentiate=TRUE` to exponentiate logistic regression coefficients.}
}
\description{
Extract model gof A mostly internal function with some potential uses
outside.
}
