% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gof.R
\name{get_gof}
\alias{get_gof}
\title{Extract model gof A mostly internal function with some potential uses
outside.}
\usage{
get_gof(model, vcov_type = NULL, ...)
}
\arguments{
\item{model}{a single model object}

\item{vcov_type}{string vcov type to add at the bottom of the table}

\item{...}{all other arguments are passed through to the extractor and
table-making functions (by default \code{broom::tidy} and \code{kableExtra::kbl}, but
this can be customized). This allows users to pass arguments directly to
\code{modelsummary} in order to affect the behavior of other functions behind the
scenes. For example,
\itemize{
\item \code{metrics="none"}, \code{metrics="all"}, or \code{metrics=c("R2", "RMSE")} to select the goodness-of-fit extracted by the \code{performance} package (must have set \code{options(modelsummary_get="easystats")} first). This can be useful for some models when statistics take a long time to compute. See \code{?performance::performance}
}}
}
\description{
Extract model gof A mostly internal function with some potential uses
outside.
}
