% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gof.R
\name{get_gof}
\alias{get_gof}
\title{Extract model gof A mostly internal function with some potential uses
outside.}
\usage{
get_gof(model, vcov_type = NULL, ...)
}
\arguments{
\item{model}{a single model object}

\item{vcov_type}{string vcov type to add at the bottom of the table}

\item{...}{all other arguments are passed through to the extractor and
table-making functions. This allows users to pass arguments directly to
\code{modelsummary} in order to affect the behavior of other functions behind
the scenes. Examples include:
\itemize{
\item \code{broom::tidy(exponentiate=TRUE)} to exponentiate logistic regression. Please see the \code{modelsummary} vignette on the package website for important technical notes on this topic.
\item \code{performance::model_performance(metrics="RMSE")} to select goodness-of-fit statistics to extract using the \code{performance} package (must have set \code{options(modelsummary_get="easystats")} first).
}}
}
\description{
Extract model gof A mostly internal function with some potential uses
outside.
}
