% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gof.R
\name{get_gof_broom}
\alias{get_gof_broom}
\title{Extract goodness-of-fit statistics from a single model using the
\code{broom} package or another package with package which supplies a
method for the \code{generics::glance} generic.}
\usage{
get_gof_broom(model, ...)
}
\description{
Extract goodness-of-fit statistics from a single model using the
\code{broom} package or another package with package which supplies a
method for the \code{generics::glance} generic.
}
\keyword{internal}
