% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsummary_list.R
\name{glance.modelsummary_list}
\alias{glance.modelsummary_list}
\title{internal glance method}
\usage{
\method{glance}{modelsummary_list}(x, ...)
}
\description{
internal glance method
}
\keyword{internal}
