% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_custom.R
\name{glance_custom}
\alias{glance_custom}
\title{Extract custom information from a model object and turn it into a tidy
data.frame or tibble with a single row.}
\usage{
glance_custom(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{ellipsis}
}
\description{
To customize the output of a model of class \code{lm}, you can define a new
method called \code{glance_custom.lm} which returns a one-row data.frame.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("glance")}
}

