% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_map.R
\docType{data}
\name{gof_map}
\alias{gof_map}
\title{Data.frame used to clean up and format goodness-of-fit statistics}
\format{data.frame with 4 columns of character data: raw, clean, fmt, omit}
\usage{
gof_map
}
\description{
Data.frame used to clean up and format goodness-of-fit statistics
}
\keyword{datasets}
