% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsummary_wide.R
\name{modelsummary_wide}
\alias{modelsummary_wide}
\title{Superseded function}
\usage{
modelsummary_wide(...)
}
\description{
This function is superseded by the \code{shape} argument of the \code{modelsummary} function.
}
\keyword{internal}
