% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{Percent}
\alias{Percent}
\title{Pseudo-function to compute a statistic relative to a reference set.}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tables}{\code{\link[tables]{Percent}}}
}}

