% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{PlusMinus}
\alias{PlusMinus}
\title{Generate 'x +/- y' terms in table.}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tables}{\code{\link[tables]{PlusMinus}}}
}}

