% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary.R
\name{sanity_check}
\alias{sanity_check}
\title{Internal function to check the sanity of various user inputs}
\usage{
sanity_check(gof_map = NULL)
}
\description{
Internal function to check the sanity of various user inputs
}
