% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_statistic_override.R
\name{statistic_override_lmtest}
\alias{statistic_override_lmtest}
\title{Use the lmtest::coeftest function to extract uncertainty estimates}
\usage{
statistic_override_lmtest(model, statistic_override)
}
\arguments{
\item{model}{object type with an available `tidy` method.}

\item{statistic_override}{manually override the uncertainy estimates. This 
argument accepts three types of input:
\itemize{
  \item a function or list of functions of length(models) which produce variance-covariance matrices with row and column names equal to the names of your coefficient estimates. For example, `R` supplies the `vcov` function, and the `sandwich` package supplies `vcovHC`, `vcovHAC`, etc.
  \item a list of length(models) variance-covariance matrices with row and column names equal to the names of your coefficient estimates.
  \item a list of length(models) numeric vectors with names equal to the names of your coefficient estimates.
}}
}
\value{
tibble
}
\description{
Use the lmtest::coeftest function to extract uncertainty estimates
}
