% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom.R
\name{tidy_custom}
\alias{tidy_custom}
\title{Extract custom information from a model object and turn it into a tidy
data.frame or tibble}
\usage{
tidy_custom(x)
}
\arguments{
\item{x}{An object to be converted into a tidy data.frame or tibble.}
}
\value{
A data.frame or tibble with information about model components.
}
\description{
To customize the output of a model of class `lm`, you can define a method
called `tidy_custom.lm` which returns a data.frame with a column called
"term", and the other columns you want to use as "estimate" or "statistic"
in your `modelsummary()` call. The output of this method must be similar to
the result of `tidy(model)`.
}
